app/requirements.txt
kafka==1.3.5
paho-mqtt==1.3.1
PyMySQL==0.9.2
redis==2.10.6
JPype1==0.6.3
numpy==1.15.1
